CLI retrieve command
